requirements.txt
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.0
tqdm>=4.64.0
pyyaml>=6.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
